requirements.txt
Package             Version
------------------- ---------
asgiref             3.4.1
beautifulsoup4      4.10.0
certifi             2021.5.30
cffi                1.15.0
charset-normalizer  2.0.5
coreapi             2.3.3
coreschema          0.0.4
cryptography        36.0.0
defusedxml          0.7.1
Django              3.2.7
django-allauth      0.46.0
django-filter       2.4.0
django-js-asset     1.2.2
django-mptt         0.13.4
django-rest-auth    0.9.5
django-rest-swagger 2.2.0
djangorestframework 3.12.4
drf-yasg            1.20.0
idna                3.2
inflection          0.5.1
itypes              1.2.0
Jinja2              3.0.1
lxml                4.6.4
Markdown            3.3.4
MarkupSafe          2.0.1
oauthlib            3.1.1
openapi-codec       1.3.2
packaging           21.0
Pillow              8.3.2
pip                 21.3.1
psycopg2            2.9.1
pycparser           2.21
PyJWT               2.3.0
pyparsing           2.4.7
python3-openid      3.2.0
pytz                2021.1
requests            2.26.0
requests-oauthlib   1.3.0
ruamel.yaml         0.17.16
ruamel.yaml.clib    0.2.6
setuptools          57.4.0
simplejson          3.17.5
six                 1.16.0
soupsieve           2.3.1
sqlparse            0.4.2
tk                  0.1.0
uritemplate         3.0.1
urllib3             1.26.6
wheel               0.37.0
